+++ npm-packages/meteor-installer/config.js
@@ -31,6 +31,10 @@ if (isWindows() && !localAppData) {
   throw new Error('LOCALAPPDATA env var is not set.');
 }
 
+const shouldSetupExecPath = () => {
+  return !process.env.npm_config_ignore_meteor_setup_exec_path;
+}
+
 const meteorLocalFolder = '.meteor';
 const meteorPath = path.resolve(rootPath, meteorLocalFolder);
 
@@ -45,5 +49,6 @@ module.exports = {
   isWindows,
   isMac,
   isRoot,
-  isSudo
+  isSudo,
+  shouldSetupExecPath,
 };